Final commit of main functionality for the application. Thorough testing to be conducted before building the application for release.
Reverted InitialRouteName in App.js to the Home.js page. Uncommented the dynamic model various and removed hard coded variable used for testing. Added some minor styling adjustments to Home.js and UploadProcessing.js. Added VWPolo marker images to the project.

FordFocusAR.js
Moved OnAnchorFound function to the top of the file for easier access. Added in comments regarding the unstability of the transformBehaviours attribute in the ViroText component. It would cause the application to completely crash on loading of the 3D objects within the ARImageMarker components. To resolve this somewhat, the visible attribute has been added so that onAnchorFound the ViroText component is no longer visible. This seems to have worked for now. Adjusted scaling of the 3D objects.

VWPoloAR.js
Implemented the same changes as in FordFocusAR.js, with the addition of file path changes due to folder restructuring, as well as changes to the 3D object positioning.